backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2
openai==1.99.6
httpx==0.28.1
requests==2.32.3
